feat():level_2
-会议模块的实体类和dao层

+++ level2/src/main/java/com/yealink/level1/bean/Staff.java
@@ -32,7 +32,7 @@ public class Staff {
 
     private String email;
 
-    private int gender;
+    private byte gender;
 
     private long createTime;
