+++ level2/src/main/java/com/yealink/level1/bean/request/PersonalRequest.java
@@ -44,7 +44,7 @@ public class PersonalRequest {
 
     private String name;
 
-    private int gender;
+    private byte gender;
 
     private String email;
